LSP context: move debug logging behind env variables
